- Starting to add RC - Collect together rows, columns, blocks, they're the same, functionally. The grid (or game?) can hold them,
And either their coordinates, or some kind of circular reference between a cell, and the groups it is a member of.

In that way, set_value can ask all groups it is a part of to validate

Seeing now that RC didnt work as I envisioned it. Does the grid also need the same RC? Then how can I enforce that only the grid changes the data?!